// client-app/src/app/models/accountant.ts
export interface Accountant {
  id: string;
    emri: string;
    datelindja: Date | null;
    vendbanimi: string;
    nrKontaktues: string;
    mbiemri: string,
    userName: string,
    email: string,
    password: string,
    gjinia: string
    token:string | null;
     kualifikimi: string;
    specializimi: string;
  }